# openapi.yaml
openapi: 3.1.0
info:
  title: Canopy Events API
  version: 1.0.0
  summary: Events among friends, for the web page and the iOS and Android apps.
  description: |
    The JSON API behind `events.canopysf.com`. The web pages use it with the
    `canopy_session` cookie; apps use it with `Authorization: Bearer <token>`,
    where the token is a `canopy_session` value. `docs/api.md` in the repo is
    the narrative guide: auth, the RSVP state machine, the visibility rules,
    pagination and errors, with worked examples.

    **Conventions.** JSON in and out. Times are ISO 8601 UTC strings
    (`2026-10-31T03:00:00.000Z`); input times need `Z` or an offset. Ids are
    strings. Errors are `{"error": "<a sentence>", "reason": "<snake_case>"}`
    with the right status: branch on `reason`, show `error` if you like.
    Lists take `?cursor=&limit=` and answer with `nextCursor` (null at the
    end). Every answer is `Cache-Control: no-store`.

    **Other people** are only ever `Person`: id, first and last name, short
    name and photo. Nobody's email, phone, Instagram, Venmo or Cash App is
    ever in an answer, except your own, from `GET /api/v1/me`.
  license:
    name: Private
    identifier: LicenseRef-Proprietary
servers:
  - url: https://events.canopysf.com
    description: Production
  - url: http://localhost:3001
    description: Local development
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/events:
    post:
      tags: [Events]
      operationId: createEvent
      summary: Make an event
      description: |
        Verified people only: an unverified (quick) account gets a 403 with
        `reason: email_unverified` and a `verify` link. You become its
        creator and only host. Limited to 20 a day per person (429
        `rate_limited`). Fields left out take their defaults; fields this
        server doesn't know are ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
            examples:
              dinner:
                value:
                  title: Rooftop dinner
                  description: Bring a jacket.
                  startsAt: '2026-10-31T19:30:00-07:00'
                  endsAt: '2026-10-31T23:00:00-07:00'
                  timeZone: America/Los_Angeles
                  locationName: "Ana's place"
                  locationAddress: 1 Market St, San Francisco
                  guestListVisibility: responded
      responses:
        '201':
          description: The new event.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      operationId: getEvent
      summary: One event
      description: |
        Anyone with the link, signed in or not. Signed out, you get the
        public details for a link preview or a "sign in to answer" page:
        no guest names, no street address (`locationAddress` is null and
        `locationAddressHidden` true), `viewer` null and no `friendsGoing`.
        Signed in, you get the address, your own part (`viewer`) and which
        of your friends are going.
      security:
        - {}
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: The event, as you see it.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Edit, cancel or un-cancel an event
      description: |
        Hosts only (403 `hosts_only`): the creator and co-hosts. Send only
        the fields to change; null or "" clears an optional one. `status:
        cancelled` cancels it (it then takes no answers or invitations),
        `status: active` takes that back; only the creator may (403
        `creator_only`). There's no deleting: a cancelled event keeps its
        link and its guest list.

        Lowering `guestsAllowed` leaves answers that already bring more as
        they are, flagged with `guestsOverLimit`; their next change has to
        fit.

        `capacity` caps people going plus their plus-ones (null: no cap).
        Raising or clearing it promotes from the waitlist at once.
        Lowering it under how many are going bumps nobody; new `going`
        answers wait until there's room.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventPatch' }
            examples:
              cancel:
                value: { status: cancelled }
              move:
                value: { startsAt: '2026-11-01T19:30:00-07:00', endsAt: null }
              plusOnes:
                value: { guestsAllowed: 2 }
      responses:
        '200':
          description: The event after the change.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /api/v1/events/{id}/cohosts:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [Hosts]
      operationId: addCohost
      summary: Make someone a co-host
      description: |
        The event's creator only (403 `creator_only`), and they have to be
        verified themselves. Co-hosts can edit the event and invite
        people; only the creator manages co-hosts, cancels, and makes a
        new link. At most 10 co-hosts (409 `too_many_cohosts`).

        **The co-host has to be verified too.** Canopy accounts doesn't
        tell sites whether someone else is, so it goes by what events saw
        the last time they used it: someone who hasn't opened Canopy
        Events signed in with a verified email gets a 403
        `email_unverified` with **no** `verify` link (that link is only
        ever for the caller's own email). They can open any event link,
        then try again.

        Becoming a co-host replaces any answer or invitation they had
        (hosts don't answer their own events), plus-ones included. Adding
        someone who's already a co-host changes nothing and answers 200.
        409 for a cancelled event or one that's over, and `is_creator`
        for yourself. 404 `person_not_found` for an id with no account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [personId]
              properties:
                personId: { type: string, description: A Canopy person id. }
            examples:
              ben:
                value: { personId: 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f }
      responses:
        '200':
          description: The event, with them in `hosts` as `cohost`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}/cohosts/{personId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
      - $ref: '#/components/parameters/PersonId'
    delete:
      tags: [Hosts]
      operationId: removeCohost
      summary: Take a co-host off, or step down
      description: |
        The creator takes off any co-host; a co-host can take themselves
        off (anyone else: 403 `creator_only`). They're left **invited**,
        so the event stays in their invitations and they can answer like
        any guest. 404 `not_cohost` if they aren't one.
      responses:
        '200':
          description: The event after.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        For apps. The token is a `canopy_session` value (43 characters,
        base64url), from the account service's native sign-in (still to be
        built). A request with a bearer header is signed in by it alone,
        never by a cookie, skips the Origin check, and never gets a
        `Set-Cookie` back. A malformed token means signed out.
    cookieAuth:
      type: apiKey
      in: cookie
      name: canopy_session
      description: |
        For the web pages. Set by `account.canopysf.com` for all of
        `canopysf.com`. Changes (POST, PUT, PATCH, DELETE) made with it
        must carry an `Origin` header for a Canopy page, or get 403
        `bad_origin`.
  parameters:
    EventId:
      name: id
      in: path
      required: true
      description: The event's id, 12 characters of base62 (the last part of its link).
      schema: { type: string, pattern: '^[0-9A-Za-z]{12}$' }
      example: 4fQ9xKpL2mZa
...
    PersonId:
      name: personId
      in: path
      required: true
      description: A Canopy person id (a UUID).
      schema: { type: string }
  responses:
    BadRequest:
      description: |
        Something in the request isn't right: `bad_json`, `bad_title`,
        `bad_starts_at`, `bad_ends_at`, `ends_before_start`,
        `bad_time_zone`, `bad_guest_list_visibility`, `bad_description`,
        `bad_location_name`, `bad_location_address`, `bad_status`,
        `bad_guests`, `too_many_guests`, `bad_guests_allowed`,
        `bad_person_ids`, `bad_person_id`, `bad_text`, `bad_capacity`,
        `bad_image`, `bad_ids`, `bad_platform`, `bad_token`, `bad_cursor`,
        `bad_limit`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'an event has to end after it starts', reason: ends_before_start }
    SignInRequired:
      description: Not signed in (`sign_in_required`), with where to send someone.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: sign in first
            reason: sign_in_required
            signIn: https://account.canopysf.com/?return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
            quickSignUp: https://account.canopysf.com/?quick=1&return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
    Forbidden:
      description: |
        Not allowed: `hosts_only`, `creator_only` (co-hosts can't do this),
        `answer_first` (posting before saying you're going or maybe),
        `not_yours` (deleting someone else's post),
        `email_unverified` (with a `verify` link when it's about your own
        email; without one when it's about someone you tried to make a
        co-host), or `bad_origin` (a change made with the cookie from a
        page that isn't Canopy's).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: verify your email first
            reason: email_unverified
            verify: https://account.canopysf.com/profile?verify=1&return=https%3A%2F%2Fevents.canopysf.com%2F
    EventNotFound:
      description: No event at that id (`event_not_found`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "there's no event at that link", reason: event_not_found }
    NotFound:
      description: '`event_not_found`, `not_invited`, `person_not_found`, `not_cohost` or `entry_not_found`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Not in this event's state: `event_cancelled`, `event_over`,
        `host_cannot_rsvp`, `already_responded`, `is_creator`,
        `too_many_cohosts`, `no_room`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'this event has been cancelled', reason: event_cancelled }
    TooLarge:
      description: The body is over 100 KB, or an image over 15 MB (`too_large`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Over a limit (`rate_limited`). Try again later.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: Canopy accounts couldn't be reached (`accounts_unreachable`). Try again in a minute.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, reason]
      properties:
        error: { type: string, description: A sentence a person can read. }
        reason: { type: string, description: What happened, as a snake_case code to branch on. }
        signIn: { type: string, format: uri, description: 'With `sign_in_required`: where to sign in and come back.' }
        quickSignUp: { type: string, format: uri, description: 'With `sign_in_required`: the quick sign-up (name, email, passkey), for someone new to Canopy.' }
        verify: { type: string, format: uri, description: 'With `email_unverified`: where to prove the email and come back.' }
...
    Person:
      description: |
        Anyone other than you, and the only shape anyone else ever has. A
        deleted account (a former member) is `firstName: "Former member"`,
        `lastName: ""`, `shortName: "Former member"`, `photoUrl: null`.
      type: object
      additionalProperties: false
      required: [id, firstName, lastName, shortName, photoUrl]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        shortName: { type: string, description: First name and last initial, e.g. "Ana L". }
        photoUrl:
          type: [string, 'null']
          format: uri
          description: |
            On account.canopysf.com. Served only with a Canopy session (the
            browser's cookie); null for no photo.
      example:
        id: 6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44
        firstName: Ana
        lastName: Lima
        shortName: Ana L
        photoUrl: https://account.canopysf.com/photo/6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44?v=1759870000000
...
    RsvpStatus:
      type: string
      enum: [invited, going, maybe, not_going, waitlisted]
      description: |
        `invited` is invited with no answer yet. `waitlisted` is a `going`
        that didn't fit the event's capacity; the server moves it to
        `going` when a spot that fits frees up. Nobody asks for it.
    Counts:
      description: |
        How many of each status. Always shown, to everyone. Hosts aren't
        counted. The top-level numbers are **people**; `guests` is the
        plus-ones those people bring, and `total` is the two together
        (people + plus-ones), for the answers that bring anyone.
      type: object
      additionalProperties: false
      required: [going, maybe, notGoing, invited, waitlisted, guests, total]
      properties:
        going: { type: integer, minimum: 0 }
        maybe: { type: integer, minimum: 0 }
        notGoing: { type: integer, minimum: 0 }
        invited: { type: integer, minimum: 0 }
        waitlisted: { type: integer, minimum: 0 }
        guests: { $ref: '#/components/schemas/GuestCounts' }
        total: { $ref: '#/components/schemas/GuestCounts' }
      example:
        going: 4
        maybe: 1
        notGoing: 1
        invited: 3
        waitlisted: 0
        guests: { going: 2, maybe: 0, waitlisted: 0 }
        total: { going: 6, maybe: 1, waitlisted: 0 }
    GuestCounts:
      type: object
      additionalProperties: false
      required: [going, maybe, waitlisted]
      properties:
        going: { type: integer, minimum: 0 }
        maybe: { type: integer, minimum: 0 }
        waitlisted: { type: integer, minimum: 0 }
    Host:
      type: object
      additionalProperties: false
      required: [person, role]
      properties:
        person: { $ref: '#/components/schemas/Person' }
        role:
          type: string
          enum: [creator, cohost]
          description: |
            The creator first, then co-hosts in the order they were added.
            Co-hosts can edit and invite; only the creator manages
            co-hosts, cancels and makes a new link.
    Rsvp:
      description: Your own place on an event.
      type: object
      additionalProperties: false
      required: [status, guests, guestsOverLimit, invited, respondedAt]
      properties:
        status: { $ref: '#/components/schemas/RsvpStatus' }
        guests: { type: integer, minimum: 0, description: Plus-ones you're bringing. }
        guestsOverLimit:
          type: boolean
          description: |
            True when the host has since lowered `guestsAllowed` below your
            `guests`. Your answer stands as it is; changing it has to fit.
        invited: { type: boolean, description: Whether a host invited you. }
        respondedAt: { type: [string, 'null'], format: date-time, description: When you last answered; null while only invited. }
    Viewer:
      description: You and this event. Null when signed out.
      type: object
      additionalProperties: false
      required: [role, rsvp, canEdit, canSeeGuestList, canPost]
      properties:
        role:
          type: [string, 'null']
          enum: [creator, cohost, null]
          description: Your part in hosting it, or null.
        rsvp:
          oneOf:
            - $ref: '#/components/schemas/Rsvp'
            - type: 'null'
        canEdit: { type: boolean }
        canSeeGuestList: { type: boolean, description: "Whether the guest list's names are shown to you. The wall follows the same rule." }
        canPost: { type: boolean, description: 'Whether you may post on the wall: hosts, and answers of going, maybe or waitlisted.' }
    FriendsGoing:
      description: |
        Which of your friends are going. `people` is the first 12, and
        empty when you can't see the guest list's names yet; `count` is
        always all of them.
      type: object
      additionalProperties: false
      required: [count, people]
      properties:
        count: { type: integer, minimum: 0 }
        people:
          type: array
          items: { $ref: '#/components/schemas/Person' }
    Event:
      type: object
      additionalProperties: false
      required: [id, url, title, description, startsAt, endsAt, timeZone, locationName, locationAddress,
                 locationAddressHidden, guestListVisibility, guestsAllowed, capacity, spotsLeft, coverImageUrl,
                 status, cancelledAt, createdAt, updatedAt,
                 hosts, counts, viewer]
      properties:
        id: { type: string, pattern: '^[0-9A-Za-z]{12}$' }
        url: { type: string, format: uri, description: The event's link, to share. }
        title: { type: string }
        description: { type: [string, 'null'] }
        startsAt: { type: string, format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        timeZone: { type: string, description: 'IANA, e.g. America/Los_Angeles: show times in it.' }
        locationName: { type: [string, 'null'] }
        locationAddress: { type: [string, 'null'], description: Null when signed out (see locationAddressHidden). }
        locationAddressHidden: { type: boolean, description: True when there's an address you'll see once signed in. }
        guestListVisibility: { type: string, enum: [everyone, responded] }
        guestsAllowed: { type: integer, minimum: 0, maximum: 10, description: Plus-ones each answer may bring. }
        capacity:
          type: [integer, 'null']
          minimum: 1
          description: The most people going, plus-ones included; null for no cap.
        spotsLeft:
          type: [integer, 'null']
          minimum: 0
          description: '`capacity` minus `counts.total.going`, never below 0; null with no cap.'
        coverImageUrl:
          type: [string, 'null']
          format: uri
          description: |
            The cover image, a JPEG, or null. Public (no sign-in needed),
            so it works in link previews; it changes with every upload,
            so cache it by URL.
        status: { type: string, enum: [active, cancelled] }
        cancelledAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        hosts:
          type: array
          items: { $ref: '#/components/schemas/Host' }
        counts: { $ref: '#/components/schemas/Counts' }
        viewer:
          oneOf:
            - $ref: '#/components/schemas/Viewer'
            - type: 'null'
        friendsGoing:
          $ref: '#/components/schemas/FriendsGoing'
          description: Only on a single event (not in lists), and only when signed in.
      example:
        id: 4fQ9xKpL2mZa
        url: https://events.canopysf.com/e/4fQ9xKpL2mZa
        title: Rooftop dinner
        description: Bring a jacket.
        startsAt: '2026-11-01T02:30:00.000Z'
        endsAt: '2026-11-01T06:00:00.000Z'
        timeZone: America/Los_Angeles
        locationName: "Ana's place"
        locationAddress: 1 Market St, San Francisco
        locationAddressHidden: false
        guestListVisibility: responded
        guestsAllowed: 1
        capacity: 20
        spotsLeft: 14
        coverImageUrl: https://events.canopysf.com/covers/Qm7Zc2pR9xTa.jpg?v=1759870000000
        status: active
        cancelledAt: null
        createdAt: '2026-10-07T20:00:00.000Z'
        updatedAt: '2026-10-07T20:00:00.000Z'
        hosts:
          - person: { id: 6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44, firstName: Ana, lastName: Lima, shortName: Ana L, photoUrl: null }
            role: creator
        counts:
          going: 4
          maybe: 1
          notGoing: 1
          invited: 3
          waitlisted: 0
          guests: { going: 2, maybe: 0, waitlisted: 0 }
          total: { going: 6, maybe: 1, waitlisted: 0 }
        viewer:
          role: null
          rsvp: { status: going, guests: 1, guestsOverLimit: false, invited: true, respondedAt: '2026-10-08T17:12:00.000Z' }
          canEdit: false
          canSeeGuestList: true
          canPost: true
        friendsGoing:
          count: 1
          people:
            - { id: 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f, firstName: Ben, lastName: Okafor, shortName: Ben O, photoUrl: null }
    EventEnvelope:
      type: object
      additionalProperties: false
      required: [event]
      properties:
        event: { $ref: '#/components/schemas/Event' }
...
    EventInput:
      description: A new event. Only title, startsAt and timeZone are required.
      type: object
      required: [title, startsAt, timeZone]
      properties:
        title: { type: string, minLength: 1, maxLength: 120, description: One line; whitespace is tidied. Longer is cut. }
        description: { type: [string, 'null'], maxLength: 5000 }
        startsAt: { type: string, format: date-time, description: 'With Z or an offset: "2026-10-31T19:30:00-07:00".' }
        endsAt: { type: [string, 'null'], format: date-time, description: After startsAt. Without one, the event counts as over 6 hours after it starts. }
        timeZone: { type: string, description: 'An IANA time zone, e.g. America/Los_Angeles.' }
        locationName: { type: [string, 'null'], maxLength: 200 }
        locationAddress: { type: [string, 'null'], maxLength: 500 }
        guestListVisibility: { type: string, enum: [everyone, responded], default: everyone }
        guestsAllowed: { type: integer, minimum: 0, maximum: 10, default: 0, description: Plus-ones each answer may bring. }
        capacity: { type: [integer, 'null'], minimum: 1, maximum: 10000, default: null, description: 'The most people going, plus-ones included. null: no cap.' }
    EventPatch:
      description: Any of the fields of EventInput, plus status. Only what's sent changes.
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: [string, 'null'], maxLength: 5000 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        timeZone: { type: string }
        locationName: { type: [string, 'null'], maxLength: 200 }
        locationAddress: { type: [string, 'null'], maxLength: 500 }
        guestListVisibility: { type: string, enum: [everyone, responded] }
        guestsAllowed: { type: integer, minimum: 0, maximum: 10, description: Lowering it keeps answers that bring more (see guestsOverLimit). }
        capacity: { type: [integer, 'null'], minimum: 1, maximum: 10000, description: "null clears it. Raising it promotes from the waitlist; lowering it bumps nobody." }
        status: { type: string, enum: [active, cancelled], description: 'cancelled cancels it; active takes that back. The creator only.' }
...